Fix texture blitting

+++ src/player.c
@@ -2,11 +2,11 @@
 
 void move(player* p, int dir) {
 	if (dir == FWD || dir == BACK) {
-		p->pos.x += cos((float)p->angle * M_PI / 180) * dir * MVMT_SPEED;
-		p->pos.z += sin((float)p->angle * M_PI / 180) * dir * MVMT_SPEED;
+		p->pos.x += (int)(cos((float)p->angle * M_PI / 180) * dir * MVMT_SPEED);
+		p->pos.z += (int)sin((float)p->angle * M_PI / 180) * dir * MVMT_SPEED;
 	} else {
-		p->pos.x += cos((float)(p->angle + 90 * (dir / 2)) * M_PI / 180) * MVMT_SPEED;
-		p->pos.z += sin((float)(p->angle + 90 * (dir / 2)) * M_PI / 180) * MVMT_SPEED;
+		p->pos.x += (int)cos((float)(p->angle + 90 * (dir / 2)) * M_PI / 180) * MVMT_SPEED;
+		p->pos.z += (int)sin((float)(p->angle + 90 * (dir / 2)) * M_PI / 180) * MVMT_SPEED;
 	}
 }
